# Conflicts:
#	.gitignore
#	Makefile
#	include/Enums/GameVersion.h
#	include/UI/PKSEFramebuffer.h
#	src/Names/AbilityNames.cpp
#	src/Names/FormNames.cpp
#	src/Names/ItemNames.cpp
#	src/UI/BackupSelectionScreen.cpp
#	src/UI/Dialogs/ItemEditDialog.cpp
#	src/UI/Dialogs/SaveConfirmDialog.cpp
#	src/UI/Dialogs/StatEditDialog.cpp
#	src/UI/Modals/PokemonDetailsModal.cpp
#	src/UI/PKSEFramebuffer.cpp
#	src/UI/Panels/BoxPokemonPanel.cpp
#	src/UI/Panels/ItemsPanel.cpp
#	src/UI/Panels/ModeSelectorPanel.cpp
#	src/UI/Panels/PartyPokemonPanel.cpp
#	src/UI/Panels/TrainerInfoPanel.cpp
#	src/UI/TitleSelectionScreen.cpp
#	src/UI/TrainerViewScreen.cpp
#	src/UI/UserSelectionScreen.cpp

+++ src/Pokemon/BaseStatsGen89.cpp
@@ -7,7 +7,11 @@ namespace Names {
     extern const char* getSpeciesName(uint16_t speciesId);
     extern const char* getItemName(uint16_t itemId);
     extern const char* getNatureName(uint8_t natureId);
-    extern const char* getAbilityName(uint8_t abilityId);
+    // uint16_t, matching the real definition in Names/AbilityNames.cpp. This used to say uint8_t,
+    // which declared a DIFFERENT overload that does not exist anywhere -- so getAbilityNameGen89()
+    // below referenced a symbol that could never resolve. It went unnoticed because nothing calls
+    // it and --gc-sections drops the function before the linker ever has to find the target.
+    extern const char* getAbilityName(uint16_t abilityId);
 }
 
 namespace Pokemon {
